Save local changes before merge

+++ src/Modules/Users/Services/user.service.js
@@ -1,6 +1,7 @@
-import { DeleteFileFromCloudinary } from "../../../Common/Services/cloudinary.service.js";
-import { User , Messages } from "../../../DB/Models/index.js";
+import mongoose from "mongoose";
 import fs from 'node:fs'
+import { DeleteFileFromCloudinary, UploadFileOnCloudinary } from "../../../Common/Services/cloudinary.service.js";
+import { User , Messages } from "../../../DB/Models/index.js";
 
 
 export const UpdateAccountService = async(req , res)=>{
@@ -9,7 +10,6 @@ export const UpdateAccountService = async(req , res)=>{
     const {_id} = req.loggedInUser
     const {firstName , lastName , email , age , gender} = req.body;
 
-    // What if the user enter an new email and it's already exists in the database
     if(email){
         const isEmailExists = await User.findOne({email})
         if(isEmailExists) return res.status(409).json({message : "Email already exists"})
@@ -30,39 +30,64 @@ export const UpdateAccountService = async(req , res)=>{
 
 
 export const DeleteAccountService = async (req, res) => {
-  // start session
   const session = await mongoose.startSession();
-    const {user: { _id },} = req.loggedInUser;
+  session.startTransaction(); // Start transaction
 
-    // start transaction
-    session.startTransaction();
+  try {
+    const {
+      user: { _id },
+    } = req.loggedInUser;
 
-    const deletedUser = await User.findByIdAndDelete({ _id }, { session });
+    const deletedUser = await User.findByIdAndDelete(_id, { session });
     if (!deletedUser) {
       await session.abortTransaction();
-      session.endSession();
       return res.status(404).json({ message: "User not found" });
     }
 
-    // unlink profile picture (locally)
-    fs.unlinkSync(deletedUser.profilePicture)
+    if (deletedUser.profilePicture && fs.existsSync(deletedUser.profilePicture)) {
+      fs.unlinkSync(deletedUser.profilePicture);
+    }
 
-    // delete by public id
-    await DeleteFileFromCloudinary(deletedUser.profilePicture.public_id)
+    if (deletedUser.profilePicture?.public_id) {
+      await DeleteFileFromCloudinary(deletedUser.profilePicture.public_id);
+    }
 
-    // delete user messages
     await Messages.deleteMany({ receiverId: _id }, { session });
 
-    // commit transaction
     await session.commitTransaction();
-
-    // end sessionnode
     session.endSession();
-    console.log("The transaction is committed");
 
-    return res.status(200).json({ message: "User deleted successfully", deletedUser }); // why we return deletedUser in the response
-  
+    return res.status(200).json({ message: "User deleted successfully" });
+  } catch (err) {
+
+    await session.abortTransaction();
+    session.endSession();
+    console.error("Error deleting user:", err);
+    return res.status(500).json({ message: "Something broke!", err: err.message });
+  }
 };
+
+
+export const UploadProfileService = async(req, res)=>{
+
+    const {_id} = req.loggedInUser
+    const {path} = req.file
+
+    const {secure_url , public_id} = await UploadFileOnCloudinary(
+        path,
+        {
+            folder: 'Nasse7a_App/Users/Profiles',
+            resource_type: 'image'
+        }
+    )
+
+    const user = await User.findByIdAndUpdate(_id,{profilePicture:{
+        secure_url,
+        public_id
+    }} , {new:true})
+
+    return res.status(200).json({message : "profile uploaded successfully" , user})
+}
 
 // List users
 export const ListUsersService = async (req, res) =>{
